% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{abort_argument_dim}
\alias{abort_argument_dim}
\title{Abort for a incorrect argument length/dimension}
\usage{
abort_argument_dim(arg, must, not = NULL)
}
\arguments{
\item{arg}{argument causing the error}

\item{must}{string specifying the desired dimension of \code{arg}}

\item{not}{current value of the \code{arg}}
}
\description{
\code{abort_argument_dim} stops execution and prints appropriate error message.
}
\keyword{internal}
