% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{abort_bad_argument}
\alias{abort_bad_argument}
\title{Abort for a NULL or type-mismatched argument}
\usage{
abort_bad_argument(arg, must, not = NULL)
}
\arguments{
\item{arg}{argument causing the error}

\item{must}{string specifying the desired type of \code{arg}}

\item{not}{current value of the \code{arg}}
}
\description{
\code{abort_bad_argument} stops execution and prints appropriate error message.
}
\keyword{internal}
