% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{abort_infeasible_argument}
\alias{abort_infeasible_argument}
\title{Abort for a infeasible argument}
\usage{
abort_infeasible_argument(arg, must, not)
}
\arguments{
\item{arg}{argument causing the error}

\item{must}{string specifying the infeasibility of \code{arg}}
}
\description{
\code{abort_infeasible_argument} stops execution and prints appropriate error message.
}
\keyword{internal}
