% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_wgn_output_statistics.R
\name{calc_rep_stats}
\alias{calc_rep_stats}
\title{Calculate statistics from all the replicates of a generated climate variable timeseries.}
\usage{
calc_rep_stats(wgn_data_rep, nyears = NULL)
}
\arguments{
\item{wgn_data_rep}{A list of vectors containing the replicates of input timeseries.
Each element of the list should be a named vector of numbers with names as character dates of the form
\code{\%Y-\%m-\%d}.}

\item{nyears}{number of years of data in the input, \code{wgn_data_rep}}
}
\value{
A named list of dataframes containing the calculated statistics.
}
\description{
\code{calc_rep_stats} returns a list of statistics calculated from an input list of timeseries
replicates.
}
\details{
Uses the function \code{calc_ts_stats} to calculate the statistics for each replicate. The statistics
from all the replicates are saved in a dataframe corresponding to each statistic.

Similar to \code{calc_ts_stats}, the statistics calculated are:
\itemize{
  \item annual maximum values (name: \emph{ann_max})
  \item annual total values (name: \emph{ann_tot})
  \item monthly mean climatology (name: \emph{mon_mean})
  \item monthly standard deviation(name: \emph{mon_sd})
  \item correlation of subsequent non-zero days by month (name: \emph{cor_wetdays_daily}),
  \item variance of non-zero days by month (name: \emph{var_wetdays_daily}),
  \item proportion of dry (zero rainfall) days within a month (name: \emph{pdry}).
}
}
