% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_wgn_output_statistics.R
\name{calc_ts_stats}
\alias{calc_ts_stats}
\title{Calculate statistics from a climate variable timeseries.}
\usage{
calc_ts_stats(wgn_data_vec)
}
\arguments{
\item{wgn_data_vec}{A named vector of numbers with names as character dates of the form
\code{\%Y-\%m-\%d}; the input daily timeseries.}
}
\value{
A named list of vectors containing the calculated statistics.
}
\description{
\code{calc_ts_stats} returns a list of statistics calculated from an input timeseries.
}
\details{
The statistics calculated are:
\itemize{
  \item annual maximum values (name: \emph{ann_max})
  \item annual total values (name: \emph{ann_tot})
  \item monthly mean climatology (name: \emph{mon_mean})
  \item monthly standard deviation(name: \emph{mon_sd})
  \item correlation of subsequent non-zero days by month (name: \emph{cor_wetdays_daily}),
  \item variance of non-zero days by month (name: \emph{var_wetdays_daily}),
  \item proportion of dry (zero rainfall) days within a month (name: \emph{pdry}).
}
}
