% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgn_latent_rain.R
\name{gen_WGN_ts}
\alias{gen_WGN_ts}
\title{Generate replicates of timeseries of a climate variable using a monthly latent variable model.}
\usage{
gen_WGN_ts(nyears = 122, replicates = 20)
}
\arguments{
\item{nyears}{number of years of data to be generated.}

\item{replicates}{number of replicates of the timeseries to be generated.}
}
\value{
A list of length \code{replicates} containing the replicates of \code{nyears} of climate variable timeseries.
}
\description{
\code{gen_WGN_ts} returns replicates of timeseries of a climate variable using the function \code{wgn_latent_model}.
}
\details{
The parameters of the weather generator are stored in toyWGN_param.env. The user may use the function \code{modify_namelist}
to modfiy the default parameters of the weather generator. This WGN model uses the parameters:
\itemize{
  \item mu vector of numbers of length 12; mean of the latent variable \emph{X} for each month.
  \item sigma vector of numbers of length 12; standard deviaion of the latent variable \emph{X} for each month.
  \item alpha vector of numbers of length 12; lag-1 autocorrelation coefficient of the latent variable \emph{X} for each month.
  \item lambda vector of numbers of length 12; exponent of the latent variable \emph{X} for each month.
  }
}
\examples{
wgn_data_rep <- gen_WGN_ts(nyears = 10, replicates = 5)
}
\seealso{
\code{write_namelist}, \code{modify_namelist}
}
