% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgn_latent_rain.R
\name{get_mon_indices}
\alias{get_mon_indices}
\title{Get indices correponding to the 12 months for a vector of dates.}
\usage{
get_mon_indices(date_vec)
}
\arguments{
\item{date_vec}{vector of dates of class \code{Date} or \code{character}}
}
\value{
list of indices of dates corresponding to each month
}
\description{
\code{get_mon_indices} returns a 12 element list of vectors containing the indices of dates the corresponding to each month
in the input vector argument.
}
\keyword{internal}
