% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parameters.R
\name{modify_namelist}
\alias{modify_namelist}
\title{Modify parameters of the weather generator}
\usage{
modify_namelist(user_file = "user_namelist.json")
}
\arguments{
\item{user_file}{A JSON file; containing name value pairs of the user input parameters}
}
\value{
The function does not return anything. It modifies the parameter values inside the parameter environment in place.
}
\description{
\code{modify_namelist} modifes the values of the WGN parameters as per user input
}
