% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_data.R
\docType{data}
\name{obs_ann_rain_max}
\alias{obs_ann_rain_max}
\title{Annual Maximum Rainfall}
\format{Numeric vector of length 122}
\source{
onk.silo.daily.csv
}
\usage{
obs_ann_rain_max
}
\description{
Example of observed annual maximum rainfall statistic calculated from data
    at station no. 23700 from 01 Jan 1889 to 31 Dec 2010 (122 years)
}
\examples{
  obs_ann_rain_max
}
\keyword{datasets}
