% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_data.R
\docType{data}
\name{obs_mon_rain_sd}
\alias{obs_mon_rain_sd}
\title{Standard Deviation of Monthly Rainfall}
\format{Named numeric vector of length 12. The names of the vector are the months
    from Jan to Dec.}
\source{
onk.silo.daily.csv
}
\usage{
obs_mon_rain_sd
}
\description{
Example of observed monthly rainfall standard deviations calculated from data
    at station no. 23700 from 01 Jan 1889 to 31 Dec 2010
}
\examples{
  obs_mon_rain_sd
}
\keyword{datasets}
