% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_data.R
\docType{data}
\name{obs_pdry}
\alias{obs_pdry}
\title{Proportion of Dry Days}
\format{Named numeric vector of length 12. The names of the vector are the months
    from Jan to Dec.}
\source{
onk.silo.daily.csv
}
\usage{
obs_pdry
}
\description{
Example of observed proportion of dry (zero-rainfall) days calculated from data
    at station no. 23700 from 01 Jan 1889 to 31 Dec 2010
}
\examples{
  obs_pdry
}
\keyword{datasets}
