% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_data.R
\docType{data}
\name{obs_var_wetdays_daily}
\alias{obs_var_wetdays_daily}
\title{Variance of Wet Day Rainfall}
\format{Named numeric vector of length 12. The names of the vector are the months
    from Jan to Dec.}
\source{
onk.silo.daily.csv
}
\usage{
obs_var_wetdays_daily
}
\description{
Example of observed variance of wet day rainfall by month calculated from data
    at station no. 23700 from 01 Jan 1889 to 31 Dec 2010
}
\examples{
  obs_var_wetdays_daily
}
\keyword{datasets}
