% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toyWGN.R
\docType{package}
\name{toyWGN}
\alias{toyWGN}
\title{toyWGN: A package to generate stochastic timeseries of climate variables, compute timeseries statistics, and create plots to compare the statistics of timeseries with observed data}
\description{
The toyWGN package provides three categories of important functions:
type1, type2 and type3.
}
\section{toyWGN functions}{

The toyWGN functions ...
}

