% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgn_latent_rain.R
\name{wgn_latent_model}
\alias{wgn_latent_model}
\title{Generate timeseries of a climate variable using a latent-variable model.}
\usage{
wgn_latent_model(mu, sigma, alpha, lambda, tot_days)
}
\arguments{
\item{mu}{A number; mean of the latent variable \emph{X}.}

\item{sigma}{A number; standard deviaion of the latent variable \emph{X}.}

\item{alpha}{A number; lag-1 autocorrelation coefficient of the latent variable \emph{X}.}

\item{lambda}{A number; exponent of the latent variable \emph{X}.}

\item{tot_days}{Number of values to return.}
}
\value{
The climate timeseries of length \code{tot_days}.
}
\description{
\code{wgn_latent_model} returns timeseries of a climate variable using the values present in its arguments.
}
\details{
The latent-variable model used by this function is of the form
\deqn{X_t = \alpha(X_{t-1} - \mu) + \epsilon_t}
\deqn{\epsilon_t ~ N(0, \sigma^2)}
\deqn{R = X^\lambda if X > 0}
}
\examples{
wgn_latent_model(-2, 1, 0.8, 1.2, 100)
}
