% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parameters.R
\name{write_namelist}
\alias{write_namelist}
\title{Write parameters of the weather generator to file}
\usage{
write_namelist(out_file = "output_namelist")
}
\arguments{
\item{output_file}{the name of the JSON file to be output}
}
\value{
The current value of all the WGN parameters in the JSON file
}
\description{
\code{output_namelist} writes the current value of the WGN parameters to a JSON file
}
